New Gun
I made a few structure changes:
Batches are now created in RenderSystem instead of GameWorld.
Environment is now created in RenderSystem instead of GameWorld

The new gun has been added. 
This new gun moves through the map with the player,
it also has a firing animation, which should set up the addition of other types of animations.